Hotfix and new functionality
